Continue splitting currency module into submodules
